requirements.txt
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
audioop-lts==0.2.1
beautifulsoup4==4.12.3
binaryornot==0.4.4
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
comtypes==1.4.9
cookiecutter==2.6.0
debugpy==1.8.11
decorator==5.1.1
distlib==0.3.9
Django==5.1.5
django-cors-headers==4.6.0
django-tailwind==3.8.0
django-tinymce==4.1.0
djangorestframework==3.15.2
executing==2.1.0
filelock==3.17.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.7.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.2.2
opencv-python==4.11.0.86
packaging==24.2
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
pure_eval==0.2.3
PyAudio==0.2.14
Pygments==2.19.1
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pyttsx3==2.98
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
rich==13.9.4
six==1.17.0
soupsieve==2.6
SpeechRecognition==3.14.1
sqlparse==0.5.3
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
tailwind==3.1.5b0
text-unidecode==1.3
tornado==6.4.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.29.1
wcwidth==0.2.13
wikipedia==1.4.0
